[example-apps] Extract platform independent server init and datamodel
