+++ examples/common/chip-app-server/BUILD.gn
@@ -1,0 +1,37 @@
+# Copyright (c) 2020 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import("//build_overrides/chip.gni")
+
+config("includes") {
+  include_dirs = [ ".", "include" ]
+}
+
+static_library("chip-app-server") {
+  output_name = "chip-app-server"
+
+  sources = [
+    "server.cpp",
+  ]
+
+  deps = [
+    "${chip_root}/src/lib",
+  ]
+
+  public_configs = [
+    ":includes",
+  ]
+
+  output_dir = root_out_dir
+}
