+++ examples/lock-app/lock-common/include/BoltLockManager.h
@@ -22,7 +22,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-#include "AppEvent.h"
+struct AppEvent;
 
 class BoltLockManager
 {
