+++ examples/lock-app/nrf5/BUILD.gn
@@ -68,9 +68,9 @@ nrf5_executable("lock_app") {
     "${chip_root}/src/app/util/process-cluster-message.c",
     "${chip_root}/src/app/util/process-global-message.c",
     "${chip_root}/src/app/util/util.c",
-    "${nrf5_platform_dir}/app/Server.cpp",
+    "${nrf5_platform_dir}/app/service.cpp",
     "${nrf5_platform_dir}/app/chipinit.cpp",
-    "${nrf5_platform_dir}/app/include/Server.h",
+    "${nrf5_platform_dir}/app/include/service.h",
     "${nrf5_platform_dir}/app/include/chipinit.h",
     "${nrf5_platform_dir}/app/support/CXXExceptionStubs.cpp",
     "${nrf5_platform_dir}/app/support/nRF5Sbrk.c",
@@ -78,19 +78,18 @@ nrf5_executable("lock_app") {
     "${nrf5_platform_dir}/util/include/LEDWidget.h",
     "main/AppTask.cpp",
     "main/BoltLockManager.cpp",
-    "main/DataModelHandler.cpp",
     "main/gen/call-command-handler.c",
     "main/gen/callback-stub.c",
     "main/gen/znet-bookkeeping.c",
     "main/include/AppEvent.h",
     "main/include/AppTask.h",
-    "main/include/BoltLockManager.h",
-    "main/include/DataModelHandler.h",
     "main/main.cpp",
   ]
 
   deps = [
     ":sdk",
+    "${chip_root}/examples/common/chip-app-server",
+    "${chip_root}/examples/lock-app/lock-common",
     "${chip_root}/src/lib",
     "${chip_root}/src/setup_payload",
     "${chip_root}/third_party/openthread/platforms/nrf528xx:libnordicsemi_nrf52840_radio_driver_softdevice",
