+++ src/lib/support/logging/CHIPLogging.cpp
@@ -88,6 +88,7 @@ static const char ModuleNames[] = "-\0\0" // None
                                   "SH\0"  // Shell
                                   "DL\0"  // DeviceLayer
                                   "SPL"   // SetupPayload
+                                  "APP"   // App
     ;
 
 #define ModuleNamesCount ((sizeof(ModuleNames) - 1) / ChipLoggingModuleNameLen)
